+++ src/core/config/ConfigType.ts
