+++ src/modules/EntityCreateModule.ts
@@ -1,7 +1,7 @@
 import { EventHandler } from "../core/handler/EventHandler";
 import { Module } from "../core/Module";
 import { Violation } from "../core/Violation";
-import { Config } from "../config/Config";
+import { Config } from "../core/config/Config";
 import { Utility } from "../util/Utility";
 
 export class EntityCreateModule extends Module {
